refactor: Staff and Customer extends from User
