+++ src/main/java/com/eequalsmc2/IoTBay_Final/model/Staff.java
@@ -3,83 +3,11 @@ package com.eequalsmc2.IoTBay_Final.model;
 import java.io.Serializable;
 import java.util.Date;
 
-public class Staff implements Serializable {
-    private int id;
-    private String email;
-    private String password;
-    private String firstName;
-    private String lastName;
-    private String phone;
-    private String gender;
-    private Date dob;
+public class Staff extends User implements Serializable {
     private int privilege;
     private String position;
 
     public Staff() {}
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public String getPhone() {
-        return phone;
-    }
-
-    public void setPhone(String phone) {
-        this.phone = phone;
-    }
-
-    public String getGender() {
-        return gender;
-    }
-
-    public void setGender(String gender) {
-        this.gender = gender;
-    }
-
-    public Date getDob() {
-        return dob;
-    }
-
-    public void setDob(Date dob) {
-        this.dob = dob;
-    }
 
     public int getPrivilege() {
         return privilege;
@@ -100,16 +28,16 @@ public class Staff implements Serializable {
     @Override
     public String toString() {
         return "Staff{" +
-                "id=" + id +
+                "privilege=" + privilege +
+                ", position='" + position + '\'' +
+                ", id=" + id +
                 ", email='" + email + '\'' +
                 ", password='" + password + '\'' +
                 ", firstName='" + firstName + '\'' +
                 ", lastName='" + lastName + '\'' +
-                ", phone='" + phone + '\'' +
                 ", gender='" + gender + '\'' +
                 ", dob=" + dob +
-                ", privilege=" + privilege +
-                ", position='" + position + '\'' +
+                ", phone='" + phone + '\'' +
                 '}';
     }
 }
